Export Credentials as sharedType for integration tests
