+++ packages/realm-app-importer/src/cli.ts
@@ -22,7 +22,8 @@ import fs from "fs";
 import http from "http";
 import chalk from "chalk";
 
-import { AppImporter, Credentials } from "./AppImporter";
+import { AppImporter } from "./AppImporter";
+import { Credentials } from "./sharedTypes";
 import { AppImportServer } from "./AppImportServer";
 
 /* eslint-disable no-console */
