+++ packages/realm-app-importer/src/react-native/index.ts
@@ -16,7 +16,7 @@
 //
 ////////////////////////////////////////////////////////////////////////////
 const AppImporter = {};
-const Credentials = {};
 const AppImporterOptions = {};
 
-export { AppImporter, Credentials, AppImporterOptions };
+export { AppImporter, AppImporterOptions };
+export { Credentials } from "../sharedTypes";
